Introduced mechanism to detect whether the user is active or inactive at the moment. The states of all users are shown in the contact list. [Closes #60]
Not required fields like the phone number can be blank when updating the user profile.

The profile information shown in each comment (such as online state and number of posts) is updated correctly when the contact list is updated. [Closes #61]

Contacts in the contact list are stored in a Hash object, which allows a very fast and non-destructive update using the new Hash method Hash#nonDestructiveUpdateFromArray (references to a certain contact won't get broken on updating the contact list.

Removed Contacts.getContact.byNickname and Contacts.getContact.byFullName.