Introduced the new JSONRPC.Store class which acts as the new base class of TaskManagement, Contacts, Storage and Gallery. This made it possible to get rid much redundancy. Based on Hash, JSONRPC.Store provides both fast record handling and search and references to certain elements in a store won't get broken when updating the store.
The client exception mechanism of JSONRPC.Request correctly determines the error's message and type.

Added JSONRPC.CachedRequest class which stores every response string and detects, if a response of a certain web service method hasn't changed since the last request.

Performance improvement for task editing functionality. The task is edited as soon as the user hits the save button. If the operation fails, the old values are restored.

The same technique was applied to the task deletion and the file archivation mechanism.