A new feature has arrived: The Shoutbox
The shout box allows signed in users to discuss in real-time in a chat-like way and notifies about events such as the creation of tasks, the upload of pictures by other users etc. [Closes #78]

The feature is still highly experimental and APIs etc. may change in the near future. Nevertheless its a huge step toward the upcoming 3.0 release.